Small updates to user profile

+++ my_server/database/user_dbf.py
@@ -56,5 +56,5 @@ def get_user_scores(user_id):
 def get_user_total_votes(user_id):
     votes = db.session.query(func.sum(MovieUserScores.votes)).filter(MovieUserScores.user_id == user_id).first()[0]
     if votes:
-        return votes/2
+        return int(votes/2)
     return 0
